Run, and save to disk, a list-length corrected reporting rate model.

Code is now generic enough to take a continuous variable (usually year) plus any combination of: another continuous variable (usually list_length), a categorical variable (often a geographic context) and/or a random variable (often a taxonomic or geographic context).

make_ll_model(
  df,
  mod_file,
  force_new = FALSE,
  y,
  y_total = NULL,
  var_col = "var",
  cat_col = "cat",
  cov_col = "cov",
  random_col = NULL,
  log_cov = TRUE,
  rand_slope = TRUE,
  mod_type = "glm",
  k = if (mod_type == "gam") 6 else NULL,
  ...
)

Arguments

df

Cleaned, filtered data frame.

mod_file

Path to file where model results are/will be saved.

force_new

Logical. If mod_file exists, should the model be re-run (and the previous results overwritten)?

y

Character name of the column in df containing the y variable.

y_total

For binomial models only. Character name of the column in df containing success + failure.

var_col

Character name of column in df containing main x variable.

cat_col

Character name of column in df containing categorical data.

cov_col

Character name of column in df containing secondary x variable.

random_col

Character name of column in df containing random factor for model. Default is NULL, so no random factor in model.

log_cov

Logical. Use cov_col (T) or the log of cov_col (F) in model specification.

rand_slope

Logical. Should a random slope effect be included in the model as well as any random intercept?

mod_type

Character. Currently "glm" or "gam".

k

For mod_type = "gam" only. Integer used as k argument of mgcv::s().

...

Depending on mod_type and random_col, passed to rstanarm::stan_gamm4(), rstanarm::stan_glm() or rstanarm::stan_glmer() (e.g. chains, iter)

Value

out_file. .rds file containing named list of inputs and model results as mod.