Explore (pre- and post-) model results.

Explore (pre- and post-) model results.

explore_mod(
  taxa,
  common,
  mod_path,
  mod_type,
  resp_var = "prop",
  geo_var = "IBRA_SUB_N",
  time_var = "year",
  max_levels = 30,
  draws = 200,
  reference_year = 2000,
  recent_year = 2010,
  re_run = FALSE,
  quant_probs = c(0.05, 0.5, 0.95),
  binom_denom = "trials",
  limit_preds = TRUE
)

Arguments

taxa

Scientific name of taxa (for labelling things).

common

Common name of taxa (also used in labels).

mod_path

Path to saved model.

mod_type

Type of model (e.g. 'reporting rate', or 'occupancy')

resp_var

What is the response variable?

geo_var

Categorical variable in the model (usually biogeographic)

time_var

Continuous time variable (usually year)

max_levels

Maximum number of classes to include in categorical plots.

draws

Number of draws from posterior distribution to display in plots.

reference_year

Reference time at which to predict (and compare change)

recent_year

Time at which to predict to compare against reference_year

quant_probs

Quantiles for summarising.

limit_preds

Logical. Limit predictions to within contexts (time_var within geo_var) that exist in the data.

Value

A list with components

count_char

ggplot object. Count of levels within character variables.

y_vs_char

ggplot object. Response variable vs character variables.

count_num

ggplot object. Histogram of each numeric variable.

y_vs_num

ggplot object. Response variable vs numeric variables with geom_smooth().

pairs

GGally::ggpairs object from df.

pred

Dataframe of posterior_predict for levels of interest.

res

pred summarised by c(resp_var, time_var).